chore(api): refactor `nodes/vms/vms_types.go`: split into multiple files
Split all `Custom*` structs and marshaling code into separate files from `vms_types.go`
